+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Current
 -------
 
 * Change the buffer API to allow append to perform partial writes.
+* When a client indicates EOF, the server will close the stream via an event.
 
 Version 0.2
 -----------